# api.yaml
swagger: '2.0'
info:
  title: API микросервиса общих справочников
  version: 1.0.0
paths:
...
  /dictionaries/banks:
    get:
      tags:
        - Banks
      description: Список банков
      operationId: listBanks
      x-spring-pageable: true
      produces:
        - application/json
      parameters:
        - in: query
          name: bic
          type: string
          description: БИК банка
          required: false
        - in: query
          name: name
          type: string
          description: Наименование банка
          required: false
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/bank'
            description: Постраничный вывод результатов поиска
        403:
          description: Forbidden
    post:
      tags:
        - Banks
      description: Создание банка
      operationId: createBank
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: bank
          description: Данные о банке
          required: true
          schema:
            $ref: '#/definitions/bank'
      responses:
        201:
          description: CREATED
          schema:
            description: Созданный банк
            $ref: '#/definitions/bank'
        403:
          description: Forbidden
        409:
          description: Если указан id и банк с таким id уже существует
  /dictionaries/banks/{id}:
    get:
      tags:
        - Banks
      description: Возвращает информацию о банке
      operationId: getBank
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор банка
      responses:
        200:
          description: ОК
          schema:
            description: Информация о банке
            $ref: '#/definitions/bank'
        403:
          description: Forbidden
        404:
          description: Not Found
    put:
      tags:
        - Banks
      description: Редактирование банка
      operationId: editBank
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор банка
        - in: body
          name: bank
          description: Данные о банке
          required: true
          schema:
            $ref: '#/definitions/bank'
      responses:
        200:
          description: ОК
          schema:
            description: Сохраненный банк
            $ref: '#/definitions/bank'
        403:
          description: Forbidden
        404:
          description: Not Found
    delete:
      tags:
        - Banks
      description: Удаление банка
      operationId: deleteBank
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор банка
      responses:
        204:
          description: No Content
        403:
          description: Forbidden
        404:
          description: Not Found
  /dictionaries/banks/replication:
    post:
      tags:
        - Banks
      description: Репликация банков
      operationId: replicationBanks
      responses:
        204:
          description: No Content
        403:
          description: Forbidden
        500:
          description: Internal server error


  #***branches***
...
  /dictionaries/addresses:
    get:
      tags:
        - Addresses
      description: Список адресов
      operationId: listAddresses
      produces:
        - application/json
      parameters:
        - in: query
          name: words
          type: string
          description: Слова поиска
          required: true
        - in: query
          name: parentId
          type: integer
          format: int64
          description: Id родительской записи адреса, относительно которой производится поиск
          required: false
      responses:
        200:
          description: ОК
          schema:
            type: array
            items:
              $ref: '#/definitions/address'
            description: Список результатов поиска
        403:
          description: Forbidden
        404:
          description: Not Found
  /dictionaries/addresses/{id}/parents:
    get:
      tags:
        - Addresses
      description: Возвращает список родительских адресов
      operationId: getParents
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор адреса
      responses:
        200:
          description: ОК
          schema:
            type: array
            items:
              $ref: '#/definitions/address'
            description: Список адресов
        403:
          description: Forbidden
        404:
          description: Not Found
  /dictionaries/addresses/{id}:
    get:
      tags:
        - Addresses
      description: Возвращает информацию об адресе
      operationId: getAddress
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          required: true
          description: Идентификатор адреса
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/address'
            description: Информация об адресе
        403:
          description: Forbidden
        404:
          description: Not Found
  /dictionaries/addresses/region/list:
    get:
      tags:
        - Addresses
      description: Возвращает список адресных объектов уровня 1
      operationId: getRegionList
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/regionRs'
            description: Ответ на запрос получения списка регионов
        403:
          description: Forbidden
        404:
          description: Not Found
  /dictionaries/addresses/replication:
    post:
      tags:
        - Addresses
      description: Репликация адресов и строений
      operationId: replicationAddresses
      responses:
        204:
          description: No Content
        403:
          description: Forbidden
        409:
          description: Conflict
        500:
          description: Internal server error
  /dictionaries/addresses/update:
    post:
      tags:
        - Addresses
      description: Обновление адресов и строений
      operationId: updateAddresses
      responses:
        204:
          description: No Content
        403:
          description: Forbidden
        409:
          description: Conflict
        500:
          description: Internal server error
  /dictionaries/addresses/{id}/houses:
    get:
      tags:
        - Addresses
      description: Список строений
      operationId: listHouses
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: integer
          format: int64
          description: Id родительской записи адреса, относительно которой производится поиск строений
          required: true
        - in: query
          name: house
          type: string
          description: Номер дома
          required: false
      responses:
        200:
          description: ОК
          schema:
            type: array
            items:
              $ref: '#/definitions/house'
            description: Список результатов поиска
        403:
          description: Forbidden
        404:
          description: Not Found
...
definitions:
...
  addressLevel:
    type: string
    description: Уровень иерархии адреса
    enum:
      - region
      - area
      - city
      - locality
      - street
      - unknown
  houseType:
    type: string
    description: Тип строения
    enum:
      - unknown
      - building
      - construction
      - letter
  bank:
    type: object
    description: Банк
    properties:
      id:
        type: integer
        format: int64
        description: Id записи
      name:
        type: string
        description: Наименование
      shortName:
        type: string
        description: Краткое наименование
      bic:
        type: string
        description: БИК
      swift:
        type: string
        description: SWIFT
      description:
        type: string
        description: Описание
      corrAccount:
        type: string
        description: Корр. счет
      address:
        type: string
        description: Адрес
      changedByHand:
        type: boolean
        description: Признак изменение данных вручную
      resident:
        type: boolean
        description: Признак резидент РФ
      deleted:
        type: boolean
        description: Признак удален
...
  address:
    type: object
    description: Адрес
    properties:
      id:
        type: integer
        format: int64
        description: Id записи
      offname:
        type: string
        description: Официальное название
      shortname:
        type: string
        description: Сокращенное наименование
      level:
        $ref: '#/definitions/addressLevel'
        description: Уровень адресной иерархии
      fulladdress:
        type: string
        description: Полный адрес
      postalcode:
        type: string
        description: Почтовый адрес
  regionRs:
    type: object
    properties:
      regions:
        type: array
        items:
          $ref: '#/definitions/region'
  region:
    type: object
    description: Информация о регионе
    properties:
      id:
        type: integer
        format: int64
        description: Id записи
      name:
        type: string
        description: Наименование
      okato:
        type: string
        description: Код по классификатору ОКАТО
  house:
    type: object
    description: Строение
    properties:
      id:
        type: integer
        format: int64
        description: Id записи
      housenum:
        type: string
        description: Номер дома
      buildnum:
        type: string
        description: Номер корпуса
      strucnum:
        type: string
        description: Номер строения
      postalcode:
        type: string
        description: Почтовый индекс
      houseType:
        $ref: '#/definitions/houseType'
        description: Тип строения
...